test(security): keep checksum test linear
